I create libs to cleanup code.
